% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{bernstein.control}
\alias{bernstein.control}
\title{Title}
\usage{
bernstein.control(m = 10L)
}
\arguments{
\item{m}{An integer specifying the Bernstein order. Default is 10.}
}
\value{
A named list of control parameters for the Bernstein method.
}
\description{
Title
}
