% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRFparallel-package.R
\docType{package}
\name{CRFparallel-package}
\alias{CRFparallel}
\alias{CRFparallel-package}
\title{CRFparallel: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mbeynaerts/CRFparallel}
  \item Report bugs at \url{https://github.com/mbeynaerts/CRFpackage/issues}
}

}
\keyword{internal}
