% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{efs.control}
\alias{efs.control}
\title{Initialize control parameters of the Generelized Fellner-Schall method}
\usage{
efs.control(
  lambda.tol = 0.1,
  REML.tol = 0.01,
  ll.tol = 0.01,
  maxiter = 100L,
  lambda.max = exp(15),
  knot.margin = 0.001,
  step.control = FALSE
)
}
\arguments{
\item{lambda.tol}{Absolute tolerance for the difference in smoothing parameters between subsequent iterations. Default is \code{0.1}.}

\item{REML.tol}{Absolute tolerance for the difference in Restricted Maximum Likelihood between subsequent iterations.}

\item{ll.tol}{Absolute tolerance for the difference in Restricted Maximum Likelihood between subsequent iterations. Default is \code{0.01}.}

\item{maxiter}{Maximum number of iterations. Default is \code{100}.}

\item{lambda.max}{Maximum value of the smoothing parameters. A smoothing parameter exceedings this value will terminate the Fellner-Schall algorithm. Default is \code{exp(15)}.}

\item{knot.margin}{description}

\item{step.control}{description}
}
\value{
A named list of control parameters for the generalized Fellner-Schall method.
}
\description{
Initialize control parameters of the Generelized Fellner-Schall method
}
