% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit,CRFparallel::biSurv-method}
\alias{fit,CRFparallel::biSurv-method}
\title{Fit the cross ratio functions using splines, polynomials or Bernstein polynomials.}
\usage{
## S7 method for class <CRFparallel::biSurv>
fit(
  biSurv,
  method = c("polynomial", "spline", "bernstein"),
  control = list(),
  nleqslv.control = list(),
  fs.control = list(),
  ncores = 1,
  progress = TRUE
)
}
\arguments{
\item{biSurv}{An object of class \code{biSurv}.}

\item{method}{A character string specifying the method to be used for fitting the model. Options are 'spline' for spline-based estimation, 'polynomial' for polynomial-based estimation, and 'bernstein' for Bernstein polynomial-based estimation. Default is 'spline'.}

\item{control}{An optional list of control parameters for the \code{nleqslv} optimization algorithm. See \code{\link[=spline.control]{spline.control()}} for spline-based estimation, \code{\link[=polynomial.control]{polynomial.control()}} for polynomial-based estimation, and \code{\link[=bernstein.control]{bernstein.control()}} for Bernstein-based estimation.}

\item{nleqslv.control}{An optional list of control parameters for the \code{nleqslv} optimization algorithm. See \code{\link[=nleqslv.control]{nleqslv.control()}} for details.}

\item{fs.control}{An optional list of control parameters for the generalized Fellner-Schall method. See \code{\link[=efs.control]{efs.control()}} for details.}

\item{ncores}{An integer specifying the number of cores to use for multi-threading. Default is 1.}

\item{progress}{Logical indicating whether to print progress messages during model fitting. Default is \code{TRUE}.}
}
\value{
An object of class \code{CRFspline}, \code{CRFpoly} or \code{CRFbern} containing the fitted model parametersn and other relevant information depending on the method used for estimation.
}
\description{
Fit the cross ratio functions using splines, polynomials or Bernstein polynomials.
}
