% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{nleqslv.control}
\alias{nleqslv.control}
\title{Initialize control parameters for \code{nlesqslv} solvers}
\usage{
nleqslv.control(method = "Broyden", global = "cline", ...)
}
\arguments{
\item{method}{A character string specifying the nonlinear equation solver to use. See the \code{method} argument in \code{\link[nleqslv:nleqslv]{nleqslv::nleqslv()}} for details. Default is "Broyden".}

\item{global}{A character string specifying the global strategy. See the \code{global} argument in \code{\link[nleqslv:nleqslv]{nleqslv::nleqslv()}} for details. Default is "cline".}

\item{...}{Any other arguments passed to \code{\link[nleqslv:nleqslv]{nleqslv::nleqslv()}}. See the \code{control} argument in \code{\link[nleqslv:nleqslv]{nleqslv::nleqslv()}} for details.}
}
\value{
A named list of control parameters.
}
\description{
Initialize control parameters for \code{nlesqslv} solvers
}
