% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{polynomial.control}
\alias{polynomial.control}
\title{Initializing polynomial parameters}
\usage{
polynomial.control(degree = 3L, restrict_degree = 3L)
}
\arguments{
\item{degree}{description}

\item{restrict_degree}{description}
}
\value{
A named list of control parameters for polynomial maximum likelihood estimation.
}
\description{
Initializing polynomial parameters
}
